requirements.txt
mysql
pyyaml
